Change type int to long

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
-        SalesManager salesManager = new SalesManager(new int[]{10, 2, 400, 10});
-        int max = salesManager.max();
+        SalesManager salesManager = new SalesManager(new long[]{10, 2, 400, 10});
+        long max = salesManager.max();
         System.out.println(max);
 
     }
